+++ docs/gen-daml-docs.sh
@@ -2,14 +2,16 @@
 
 set -eou pipefail
 
+DOCS_DIR="$REPO_ROOT/docs"
+
 gen_project_docs () (
     echo "(docs) generating $1"
     cd "$REPO_ROOT/$1"
     local -a DAML_FILES
     readarray -t DAML_FILES < <(find daml -name '*.daml')
-    "${XDG_CACHE_HOME:-$HOME/.cache}/daml-build/${DAML_COMPILER_VERSION}/damlc/damlc" docs --index-template "$REPO_ROOT/cluster/images/docs/api-templates/$2-index-template.rst" "${DAML_FILES[@]}" --exclude-modules '**.Scripts.**' -f rst -o "$REPO_ROOT/cluster/images/docs/src/app_dev/api/$2"
+    "${XDG_CACHE_HOME:-$HOME/.cache}/daml-build/${DAML_COMPILER_VERSION}/damlc/damlc" docs --index-template "$DOCS_DIR/api-templates/$2-index-template.rst" "${DAML_FILES[@]}" --exclude-modules '**.Scripts.**' -f rst -o "$DOCS_DIR/src/app_dev/api/$2"
     # Workaround to fix indentation issues in rst output due to https://github.com/digital-asset/daml/issues/16956
-    find "$REPO_ROOT/cluster/images/docs/src/app_dev/api/$2" -name '*.rst' -exec sed -i -z 's!\( *\)\(Controller\\: [^\n]*map\)\n *\([^\n]*\)\n *\([^\n]*\)\n!\1\2\n\1\3\n\1\4\n!g' {} +
+    find "$DOCS_DIR/src/app_dev/api/$2" -name '*.rst' -exec sed -i -z 's!\( *\)\(Controller\\: [^\n]*map\)\n *\([^\n]*\)\n *\([^\n]*\)\n!\1\2\n\1\3\n\1\4\n!g' {} +
 )
 
 ensure_damlc_exists() {
@@ -27,11 +29,6 @@ ensure_damlc_exists() {
         popd
     fi
 }
-
-# skip calling sbt when `SKIP_DAML_BUILD` is defined
-if [[ ! -v SKIP_DAML_BUILD ]]; then
-  (cd "$REPO_ROOT"; sbt --batch damlBuild)
-fi
 
 ensure_damlc_exists