+++ docs/livepreview.sh
@@ -3,6 +3,7 @@ set -eou pipefail
 
 cd "$(dirname "${BASH_SOURCE[0]}")"
 
+(cd "$REPO_ROOT"; sbt --batch damlBuild)
 ./gen-daml-docs.sh
 
-sphinx-autobuild src target
+sphinx-autobuild src html/html